Moved component from constants file
