+++ lib/components/EmailInput.js
@@ -2,17 +2,29 @@ import React, { useState } from "react";
 
 import classnames from "classnames";
 import PropTypes from "prop-types";
+import { components } from "react-select";
 import CreatableSelect from "react-select/creatable";
 
 import {
   EMAIL_SEPARATION_REGEX,
   CUSTOM_STYLES,
-  CUSTOM_COMPONENTS,
   formatEmailInputOptions,
 } from "../constants/EmailInput";
 import { hyphenize } from "../common";
 import Typography from "./Typography";
 import Label from "./Label";
+
+const CustomControl = ({ children, ...props }) => (
+  <components.Control {...props}>
+    <div className="overflow-hidden">{children}</div>
+  </components.Control>
+);
+
+const CUSTOM_COMPONENTS = {
+  DropdownIndicator: null,
+  ClearIndicator: null,
+  Control: CustomControl,
+};
 
 const EmailInput = ({
   label = "",
