+++ lib/constants/EmailInput.js
@@ -1,6 +1,3 @@
-import React from "react";
-import { components } from "react-select";
-
 export const EMAIL_REGEX = new RegExp(
   "^[A-Z0-9._%+-]+@[A-Z0-9.-]+\\.[A-Z]{2,}$",
   "i"
@@ -29,15 +26,3 @@ export const formatEmailInputOptions = label => ({
   value: label,
   valid: EMAIL_REGEX.test(label),
 });
-
-const CustomControl = ({ children, ...props }) => (
-  <components.Control {...props}>
-    <div className="overflow-hidden">{children}</div>
-  </components.Control>
-);
-
-export const CUSTOM_COMPONENTS = {
-  DropdownIndicator: null,
-  ClearIndicator: null,
-  Control: CustomControl,
-};
